Changed to int math

+++ 3D_Renderer/ThreeTwo.java
@@ -2,17 +2,17 @@ public class ThreeTwo
 {
     public static void CalcPerspective(Point point, Point camera, Point cameraAngle)
     {
-        double[][] mat1 = {{1,0,0},
-                           {0,Math.cos(cameraAngle.getX3()), Math.sin(cameraAngle.getX3())},
-                           {0,Math.sin(cameraAngle.getX3()) * -1, Math.cos(cameraAngle.getX3())}};
-        double[][] mat2 = {{Math.cos(cameraAngle.getY3()), 0, Math.sin(cameraAngle.getY3()) * -1},
-                           {0,1,0},
-                           {Math.sin(cameraAngle.getY3()), 0, Math.cos(cameraAngle.getY3())}};
-        double[][] mat3 = {{Math.cos(cameraAngle.getZ3()), Math.sin(cameraAngle.getZ3()), 0},
-                           {Math.sin(cameraAngle.getZ3()) * -1, Math.cos(cameraAngle.getZ3()), 0},
-                           {0,0,1}};
-        double[][] PC = {{point.getX3() - camera.getX3()}, {point.getY3() - camera.getY3()}, {point.getZ3() - camera.getZ3()}};
-        double[][] xyz = MatrixMult.mult(MatrixMult.mult(MatrixMult.mult(mat1,mat2), mat3), PC);
+        int[][] mat1 = {{1,0,0},
+                        {0,(int)Math.cos(cameraAngle.getX3()), (int)Math.sin(cameraAngle.getX3())},
+                        {0,(int)Math.sin(cameraAngle.getX3()) * -1, (int)Math.cos(cameraAngle.getX3())}};
+        int[][] mat2 = {{(int)Math.cos(cameraAngle.getY3()), 0, (int)Math.sin(cameraAngle.getY3()) * -1},
+                        {0,1,0},
+                        {(int)Math.sin(cameraAngle.getY3()), 0, (int)Math.cos(cameraAngle.getY3())}};
+        int[][] mat3 = {{(int)Math.cos(cameraAngle.getZ3()), (int)Math.sin(cameraAngle.getZ3()), 0},
+                        {(int)Math.sin(cameraAngle.getZ3()) * -1, (int)Math.cos(cameraAngle.getZ3()), 0},
+                        {0,0,1}};
+        int[][] PC = {{(int)(point.getX3() - camera.getX3())}, {(int)(point.getY3() - camera.getY3())}, {(int)(point.getZ3() - camera.getZ3())}};
+        int[][] xyz = MatrixMult.mult(MatrixMult.mult(MatrixMult.mult(mat1,mat2), mat3), PC);
         point.setX3(xyz[0][0]);
         point.setY3(xyz[1][0]);
         point.setZ3(xyz[2][0]);
